move more tests to HttpServiceIntegrationTest
